pyproject.toml
[project]
name = "realestateweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
]
